Moved dummy data creation used in single script to linked scripts
