+++ tests/nominal/test_MeanResponseTransformer.py
@@ -6,9 +6,74 @@ import pytest
 import test_aide as ta
 from pandas.testing import assert_series_equal
 
-import tests.test_data as d
 import tubular
 from tubular.nominal import MeanResponseTransformer
+
+
+# Dataframe used exclusively in this testing script
+def create_MeanResponseTransformer_test_df():
+    """Create DataFrame to use MeanResponseTransformer tests that correct values are.
+
+    DataFrame column a is the response, the other columns are categorical columns
+    of types; object, category, int, float, bool.
+
+    """
+    df = pd.DataFrame(
+        {
+            "a": [1.0, 2.0, 3.0, 4.0, 5.0, 6.0],
+            "b": ["a", "b", "c", "d", "e", "f"],
+            "c": ["a", "b", "c", "d", "e", "f"],
+            "d": [1, 2, 3, 4, 5, 6],
+            "e": [1, 2, 3, 4, 5, 6.0],
+            "f": [False, False, False, True, True, True],
+            "multi_level_response": [
+                "blue",
+                "blue",
+                "yellow",
+                "yellow",
+                "green",
+                "green",
+            ],
+        },
+    )
+
+    df["c"] = df["c"].astype("category")
+
+    return df
+
+
+# Dataframe used exclusively in this testing script
+def create_MeanResponseTransformer_test_df_unseen_levels():
+    """Create DataFrame to use in MeanResponseTransformer tests that check correct values are
+    generated when using transform method on data with unseen levels.
+    DataFrame column a is the response, the other columns are categorical columns
+    of types; object, category, int, float, bool.
+
+    """
+    df = pd.DataFrame(
+        {
+            "a": [1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 3.0],
+            "b": ["a", "b", "c", "d", "e", "f", "g", "h"],
+            "c": ["a", "b", "c", "d", "e", "f", "g", "h"],
+            "d": [1, 2, 3, 4, 5, 6, 7, 8],
+            "e": [1, 2, 3, 4, 5, 6.0, 7, 8],
+            "f": [False, False, False, True, True, True, True, False],
+            "multi_level_response": [
+                "blue",
+                "blue",
+                "yellow",
+                "yellow",
+                "green",
+                "green",
+                "yellow",
+                "blue",
+            ],
+        },
+    )
+
+    df["c"] = df["c"].astype("category")
+
+    return df
 
 
 @pytest.fixture()
@@ -22,7 +87,7 @@ def learnt_mapping_dict():
         "b_green": {"a": 0.0, "b": 0.0, "c": 0.0, "d": 0.0, "e": 1.0, "f": 1.0},
     }
 
-    # c matches b, but is categorical (see test_data.create_MeanResponseTransformer_test_df)
+    # c matches b, but is categorical (see create_MeanResponseTransformer_test_df)
     c_dict = {
         "c" + suffix: b_dict["b" + suffix]
         for suffix in ["", "_blue", "_yellow", "_green"]
@@ -219,7 +284,7 @@ class Test_prior_regularisation:
 class TestFit:
     def test_super_fit_called(self, mocker):
         """Test that fit calls BaseTransformer.fit."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(columns="b")
 
@@ -241,8 +306,8 @@ class TestFit:
 
         expected_pos_args = (
             x,
-            d.create_MeanResponseTransformer_test_df(),
-            d.create_MeanResponseTransformer_test_df()["a"],
+            create_MeanResponseTransformer_test_df(),
+            create_MeanResponseTransformer_test_df()["a"],
         )
 
         assert len(expected_pos_args) == len(
@@ -265,7 +330,7 @@ class TestFit:
     )
     def test_fit_returns_self(self, level, target_column, unseen_level_handling):
         """Test fit returns self?."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(
             columns="b",
@@ -289,7 +354,7 @@ class TestFit:
     )
     def test_fit_not_changing_data(self, level, target_column, unseen_level_handling):
         """Test fit does not change X."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(
             columns="b",
@@ -300,7 +365,7 @@ class TestFit:
         x.fit(df, df[target_column])
 
         ta.equality.assert_equal_dispatch(
-            expected=d.create_MeanResponseTransformer_test_df(),
+            expected=create_MeanResponseTransformer_test_df(),
             actual=df,
             msg="Check X not changing during fit",
         )
@@ -320,7 +385,7 @@ class TestFit:
         unseen_level_handling,
     ):
         """Test that an exception is raised if nulls are present in response_column."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
         df.loc[1, target_column] = np.nan
 
         x = MeanResponseTransformer(
@@ -350,7 +415,7 @@ class TestFit:
         unseen_level_handling,
     ):
         "Test that the mapping dictionary created in fit has the correct keys and values."
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
         columns = ["b", "c"]
         x = MeanResponseTransformer(
             columns=columns,
@@ -382,7 +447,7 @@ class TestFit:
         unseen_level_handling,
     ):
         "Test that the mapping dictionary created in fit has the correct keys and values."
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
         columns = ["b", "c"]
         x = MeanResponseTransformer(
             columns=columns,
@@ -437,7 +502,7 @@ class TestFit:
         unseen_level_handling,
     ):
         "Test that the unseen_levels_encoding_dict dictionary created in fit has the correct keys and values."
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
         x = MeanResponseTransformer(
             columns=["b"],
             level=level,
@@ -502,7 +567,7 @@ class TestFit:
     def test_missing_categories_ignored(self):
         "test that where a categorical column has missing levels, these do not make it into the encoding dict"
 
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
         unobserved_value = "bla"
         df["c"] = df["c"].cat.add_categories(unobserved_value)
         target_column = "e"
@@ -521,7 +586,7 @@ class TestFitBinaryResponse:
 
     def test_learnt_values(self):
         """Test that the mean response values learnt during fit are expected."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(columns=["b", "d", "f"])
 
@@ -552,7 +617,7 @@ class TestFitBinaryResponse:
 
     def test_learnt_values_prior_no_weight(self):
         """Test that the mean response values learnt during fit are expected."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(columns=["b", "d", "f"], prior=5)
 
@@ -596,7 +661,7 @@ class TestFitBinaryResponse:
 
     def test_learnt_values_no_prior_weight(self):
         """Test that the mean response values learnt during fit are expected if a weights column is specified."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(weights_column="e", columns=["b", "d", "f"])
 
@@ -626,7 +691,7 @@ class TestFitBinaryResponse:
 
     def test_learnt_values_prior_weight(self):
         """Test that the mean response values learnt during fit are expected - when using weight and prior."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         df["weight"] = [1, 1, 1, 2, 2, 2]
 
@@ -662,7 +727,7 @@ class TestFitBinaryResponse:
     @pytest.mark.parametrize("prior", (1, 3, 5, 7, 9, 11, 100))
     def test_prior_logic(self, prior):
         "Test that for prior>0 encodings are closer to global mean than for prior=0."
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         df["weight"] = [1, 1, 1, 2, 2, 2]
 
@@ -713,7 +778,7 @@ class TestFitBinaryResponse:
     )
     def test_prior_logic_for_weights(self, low_weight, high_weight):
         "Test that for fixed prior a group with lower weight is moved closer to the global mean than one with higher weight."
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         # column f looks like [False, False, False, True, True, True]
         df["weight"] = [
@@ -780,7 +845,7 @@ class TestFitBinaryResponse:
 
     def test_weights_column_missing_error(self):
         """Test that an exception is raised if weights_column is specified but not present in data for fit."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(weights_column="z", columns=["b", "d", "f"])
 
@@ -1078,7 +1143,7 @@ class TestTransform:
 
     def test_check_is_fitted_called(self, mocker):
         """Test that BaseTransformer check_is_fitted called."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(columns="b")
 
@@ -1095,7 +1160,7 @@ class TestTransform:
             x.transform(df)
 
     def test_not_dataframe_error_raised(self):
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(columns="b")
         x.fit(df, df["a"])
@@ -1108,14 +1173,14 @@ class TestTransform:
 
     def test_super_transform_called(self, mocker):
         """Test that BaseNominalTransformer.transform called."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(columns="b")
 
         x.fit(df, df["a"])
 
         expected_call_args = {
-            0: {"args": (d.create_MeanResponseTransformer_test_df(),), "kwargs": {}},
+            0: {"args": (create_MeanResponseTransformer_test_df(),), "kwargs": {}},
         }
 
         with ta.functions.assert_function_call(
@@ -1123,13 +1188,13 @@ class TestTransform:
             tubular.nominal.BaseNominalTransformer,
             "transform",
             expected_call_args,
-            return_value=d.create_MeanResponseTransformer_test_df(),
+            return_value=create_MeanResponseTransformer_test_df(),
         ):
             x.transform(df)
 
     def test_learnt_values_not_modified(self):
         """Test that the mappings from fit are not changed in transform."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(columns="b")
 
@@ -1150,7 +1215,7 @@ class TestTransform:
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(
-            d.create_MeanResponseTransformer_test_df(),
+            create_MeanResponseTransformer_test_df(),
             expected_df_1(),
         ),
     )
@@ -1180,7 +1245,7 @@ class TestTransform:
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(
-            d.create_MeanResponseTransformer_test_df(),
+            create_MeanResponseTransformer_test_df(),
             expected_df_2(),
         ),
     )
@@ -1222,7 +1287,7 @@ class TestTransform:
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(
-            d.create_MeanResponseTransformer_test_df(),
+            create_MeanResponseTransformer_test_df(),
             expected_df_3(),
         ),
     )
@@ -1263,10 +1328,10 @@ class TestTransform:
         """Test that the output is expected from transform with a single level response with unseen levels in data with
         unseen_level_handling set to 'None', i.e., default value.
         """
-        initial_df = d.create_MeanResponseTransformer_test_df()
+        initial_df = create_MeanResponseTransformer_test_df()
         x = MeanResponseTransformer(columns=["b", "d", "f"])
         x.fit(initial_df, initial_df["a"])
-        df = d.create_MeanResponseTransformer_test_df_unseen_levels()
+        df = create_MeanResponseTransformer_test_df_unseen_levels()
         with pytest.raises(
             ValueError,
             match="MeanResponseTransformer: nulls would be introduced into column b from levels not present in mapping",
@@ -1276,7 +1341,7 @@ class TestTransform:
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(
-            d.create_MeanResponseTransformer_test_df_unseen_levels(),
+            create_MeanResponseTransformer_test_df_unseen_levels(),
             expected_df_4(),
         ),
     )
@@ -1295,7 +1360,7 @@ class TestTransform:
             unseen_level_handling="Mean",
         )
 
-        initial_df = d.create_MeanResponseTransformer_test_df()
+        initial_df = create_MeanResponseTransformer_test_df()
         x.fit(initial_df, initial_df[target])
         df_transformed = x.transform(df)
 
@@ -1311,7 +1376,7 @@ class TestTransform:
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(
-            d.create_MeanResponseTransformer_test_df_unseen_levels(),
+            create_MeanResponseTransformer_test_df_unseen_levels(),
             expected_df_5(),
         ),
     )
@@ -1330,7 +1395,7 @@ class TestTransform:
             unseen_level_handling="Median",
         )
 
-        initial_df = d.create_MeanResponseTransformer_test_df()
+        initial_df = create_MeanResponseTransformer_test_df()
         x.fit(initial_df, initial_df[target])
         df_transformed = x.transform(df)
 
@@ -1346,7 +1411,7 @@ class TestTransform:
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(
-            d.create_MeanResponseTransformer_test_df_unseen_levels(),
+            create_MeanResponseTransformer_test_df_unseen_levels(),
             expected_df_6(),
         ),
     )
@@ -1365,7 +1430,7 @@ class TestTransform:
             unseen_level_handling="Lowest",
         )
 
-        initial_df = d.create_MeanResponseTransformer_test_df()
+        initial_df = create_MeanResponseTransformer_test_df()
         x.fit(initial_df, initial_df[target])
         df_transformed = x.transform(df)
 
@@ -1381,7 +1446,7 @@ class TestTransform:
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(
-            d.create_MeanResponseTransformer_test_df_unseen_levels(),
+            create_MeanResponseTransformer_test_df_unseen_levels(),
             expected_df_7(),
         ),
     )
@@ -1400,7 +1465,7 @@ class TestTransform:
             unseen_level_handling="Highest",
         )
 
-        initial_df = d.create_MeanResponseTransformer_test_df()
+        initial_df = create_MeanResponseTransformer_test_df()
         x.fit(initial_df, initial_df[target])
         df_transformed = x.transform(df)
 
@@ -1416,7 +1481,7 @@ class TestTransform:
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(
-            d.create_MeanResponseTransformer_test_df_unseen_levels(),
+            create_MeanResponseTransformer_test_df_unseen_levels(),
             expected_df_8(),
         ),
     )
@@ -1432,7 +1497,7 @@ class TestTransform:
         target = "a"
         x = MeanResponseTransformer(columns=columns, unseen_level_handling=21.6)
 
-        initial_df = d.create_MeanResponseTransformer_test_df()
+        initial_df = create_MeanResponseTransformer_test_df()
         x.fit(initial_df, initial_df[target])
         df_transformed = x.transform(df)
 
@@ -1448,7 +1513,7 @@ class TestTransform:
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(
-            d.create_MeanResponseTransformer_test_df_unseen_levels(),
+            create_MeanResponseTransformer_test_df_unseen_levels(),
             expected_df_9(),
         ),
     )
@@ -1468,7 +1533,7 @@ class TestTransform:
         for col in expected_created_cols:
             expected[col] = expected[col].astype(x.return_type)
 
-        initial_df = d.create_MeanResponseTransformer_test_df()
+        initial_df = create_MeanResponseTransformer_test_df()
         x.fit(initial_df, initial_df["multi_level_response"])
         df_transformed = x.transform(df)
 
@@ -1481,7 +1546,7 @@ class TestTransform:
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(
-            d.create_MeanResponseTransformer_test_df_unseen_levels(),
+            create_MeanResponseTransformer_test_df_unseen_levels(),
             expected_df_10(),
         ),
     )
@@ -1489,7 +1554,7 @@ class TestTransform:
         """Test that the output is expected from transform with a multi-level response and unseen levels and all level selected."""
         columns = ["b", "f"]
         target = "multi_level_response"
-        initial_df = d.create_MeanResponseTransformer_test_df()
+        initial_df = create_MeanResponseTransformer_test_df()
         expected_created_cols = [
             prefix + "_" + suffix
             for prefix, suffix in product(columns, initial_df[target].unique().tolist())
@@ -1514,7 +1579,7 @@ class TestTransform:
 
     def test_nulls_introduced_in_transform_error(self):
         """Test that transform will raise an error if nulls are introduced."""
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         x = MeanResponseTransformer(columns=["b", "d", "f"])
 
@@ -1546,7 +1611,7 @@ class TestTransform:
     ):
         "Test that output return types are controlled by return_type param, this defaults to float32 so test float64 here"
 
-        df = d.create_MeanResponseTransformer_test_df()
+        df = create_MeanResponseTransformer_test_df()
 
         columns = ["b", "d", "f"]
         x = MeanResponseTransformer(
